move App component to folder
